refactor(scripts): decompose epic-plan-spec into phases/
Extracts the six epic-plan-spec responsibilities plus their helpers
from the monolithic epic-plan-spec.js (998 LOC) into a sibling phases/
directory under .agents/scripts/lib/orchestration/epic-plan-spec/:

  - plan-epic.js          (227 LOC) — planEpic + resolveAcceptancePersistence
  - authoring-context.js  (184 LOC) — buildAuthoringContext + resolveMemoryDir
  - run-spec-phase.js     (182 LOC) — runSpecPhase + setEpicLabel
  - spec-freshness.js     (120 LOC) — runSpecFreshnessCheck
  - drain.js              (110 LOC) — drainPendingCleanupAtBoot
  - prompts.js            ( 55 LOC) — PRD / Tech Spec / Acceptance Spec
                                       SYSTEM_PROMPT constants

The parent epic-plan-spec.js shrinks to 196 LOC and becomes a thin CLI
entry that re-exports the public API (buildAuthoringContext, planEpic,
runSpecPhase, the SYSTEM_PROMPT constants, drainPendingCleanupAtBoot,
runSpecFreshnessCheck, resolveMemoryDir, resolveAcceptancePersistence,
resolveReviewRouting, forkAndCommitEpicSnapshot, forkMainToEpic) so all
existing test imports keep working.

MAINTAINABILITY_EXCLUSIONS update: the post-split parent now scores
MI 71.241 (escomplex can parse the trimmed shape), so removed from the
allow-list. One extracted phase (spec-freshness.js) still hits the same
escomplex "this[node.expression.type] is not a function" family on its
optional-chained destructuring — added that single path with a
documented reason so the rollup isn't poisoned. Other five phase files
score MI 71–138 cleanly.

Refs #3004 #2988
