+++ .agents/scripts/epic-plan-spec.js
@@ -1,5 +1,4 @@
 #!/usr/bin/env node
-/* node:coverage ignore file */
 
 /**
  * epic-plan-spec.js — Phase 7 (spec) entry point for the split planning flow.
@@ -24,11 +23,13 @@
  * Exit codes:
  *   0 — phase complete, Epic is now `agent::review-spec`.
  *   1 — fatal error (see stderr).
+ *
+ * The phase implementations live under
+ * `lib/orchestration/epic-plan-spec/phases/`. This file is now a thin CLI
+ * entry that wires argv → phases.
  */
 
-import { mkdir, readFile, writeFile } from 'node:fs/promises';
-import * as os from 'node:os';
-import path from 'node:path';
+import { readFile } from 'node:fs/promises';
 import { parseArgs } from 'node:util';
 import {
   forkAndCommitEpicSnapshot,
@@ -42,851 +43,48 @@ import {
 // for the same reason.
 export { forkAndCommitEpicSnapshot, forkMainToEpic };
 
-import {
-  resolveFeatureRoots,
-  verifyBddRunnerPendingTag,
-} from './lib/bdd-runner-detect.js';
-import { scanBddScenarios } from './lib/bdd-scenario-scanner.js';
 import { runAsCli } from './lib/cli-utils.js';
-import { buildCodebaseSnapshot } from './lib/codebase-snapshot.js';
 import {
-  getLimits,
   PROJECT_ROOT,
   resolveConfig,
   validateOrchestrationConfig,
 } from './lib/config-resolver.js';
-import { scanMemoryFreshness } from './lib/feedback-loop/memory-freshness.js';
-import { fetchPriorFeedback } from './lib/feedback-loop/prior-feedback-fetcher.js';
-import * as gitUtils from './lib/git-utils.js';
 import { Logger, routeAllOutputToStderr, STDERR_LOGGER } from './lib/Logger.js';
 import {
-  ACCEPTANCE_NA,
-  AGENT_LABELS,
-  TYPE_LABELS,
-} from './lib/label-constants.js';
-import { buildDocsContext } from './lib/orchestration/doc-reader.js';
+  buildAuthoringContext,
+  resolveMemoryDir,
+} from './lib/orchestration/epic-plan-spec/phases/authoring-context.js';
+import { drainPendingCleanupAtBoot } from './lib/orchestration/epic-plan-spec/phases/drain.js';
 import {
-  initialize as initializePlanState,
-  PLAN_PHASES,
-  read as readPlanState,
-  setPhase as setPlanPhase,
-  write as writePlanState,
-} from './lib/orchestration/epic-plan-state-store.js';
+  planEpic,
+  resolveAcceptancePersistence,
+} from './lib/orchestration/epic-plan-spec/phases/plan-epic.js';
+import {
+  ACCEPTANCE_SPEC_SYSTEM_PROMPT,
+  PRD_SYSTEM_PROMPT,
+  TECH_SPEC_SYSTEM_PROMPT,
+} from './lib/orchestration/epic-plan-spec/phases/prompts.js';
+import { runSpecPhase } from './lib/orchestration/epic-plan-spec/phases/run-spec-phase.js';
+import { runSpecFreshnessCheck } from './lib/orchestration/epic-plan-spec/phases/spec-freshness.js';
 import { resolveReviewRouting } from './lib/orchestration/plan-review-routing.js';
-import { sweepStaleStoryWorktrees } from './lib/orchestration/plan-runner/worktree-sweep.js';
-import { applyBudget } from './lib/orchestration/planning-context-budget.js';
-import { classifyPlanningRisk } from './lib/orchestration/planning-risk.js';
-
-export { resolveReviewRouting };
-
-import { PlanningStateManager } from './lib/orchestration/planning-state-manager.js';
-import {
-  renderSpecFreshnessComment,
-  validateSpecFreshness,
-} from './lib/orchestration/spec-freshness.js';
-import { upsertStructuredComment } from './lib/orchestration/ticketing.js';
-import { cleanupPhaseTempFiles } from './lib/plan-phase-cleanup.js';
 import { createProvider } from './lib/provider-factory.js';
-import { forceDrainPendingCleanup } from './lib/worktree/lifecycle/force-drain.js';
-import { readManifest } from './lib/worktree/lifecycle/pending-cleanup.js';
 
-// ─── PRD / Tech Spec system prompts ──────────────────────────────────────────
-//
-// These are the canonical authoring prompts that ride along on the
-// `--emit-context` envelope as a backstop. The `epic-plan-spec-author` Skill
-// (`.agents/skills/core/epic-plan-spec-author/SKILL.md`) embeds the
-// authoritative copies of these strings — keep the two surfaces in sync when
-// either is edited.
+// Re-exports for stable public API: tests and external callers import these
+// from `epic-plan-spec.js`. The implementations live in `phases/`.
+export {
+  ACCEPTANCE_SPEC_SYSTEM_PROMPT,
+  buildAuthoringContext,
+  drainPendingCleanupAtBoot,
+  PRD_SYSTEM_PROMPT,
+  planEpic,
+  resolveAcceptancePersistence,
+  resolveMemoryDir,
+  resolveReviewRouting,
+  runSpecFreshnessCheck,
+  runSpecPhase,
+  TECH_SPEC_SYSTEM_PROMPT,
+};
 
-export const PRD_SYSTEM_PROMPT = `You are an expert Technical Product Manager.
-Your job is to convert a high-level Epic description into a structured Product Requirements Document (PRD).
-
-The PRD should outline:
-1. Context & Goals
-2. User Stories
-3. Acceptance Criteria
-4. Out of Scope
-
-CRITICAL REQUIREMENTS:
-- Respond ONLY with valid Markdown.
-- Do not use top-level <h1> (# ) tags. Start with ## Overview.
-- Format requirements clearly with bullet points and bold text where appropriate.`;
-
-export const TECH_SPEC_SYSTEM_PROMPT = `You are an expert Engineering Architect.
-Your job is to convert a PRD into a Technical Specification for implementation.
-
-The Tech Spec should outline:
-1. Architecture & Design
-2. Data Models (if any)
-3. API Changes (if any)
-4. Core Components
-5. Security & Privacy Considerations
-
-CRITICAL REQUIREMENTS:
-- Respond ONLY with valid Markdown.
-- Do not use top-level <h1> (# ) tags. Start with ## Technical Overview.
-- Format architectural decisions clearly with bullet points.`;
-
-export const ACCEPTANCE_SPEC_SYSTEM_PROMPT = `You are an expert Acceptance Engineer.
-Your job is to convert a PRD and a Tech Spec into a structured Acceptance Specification that drives features-first BDD authoring.
-
-The Acceptance Spec should outline:
-1. Acceptance Criteria — one row per user-visible outcome, expressed as a Markdown table with columns: AC ID | Outcome | Feature File | Scenario | Disposition
-2. Stable AC IDs — assign AC-1, AC-2, ... in document order; reuse the same ID across re-plans when an Outcome is materially unchanged so scenario tags (@ac-N) stay aligned
-3. Disposition — tag each row with one of: new | updated | unchanged
-
-CRITICAL REQUIREMENTS:
-- Respond ONLY with valid Markdown.
-- Do not use top-level <h1> (# ) tags. Start with ## Acceptance Criteria.
-- Every AC row MUST have a stable AC ID of the form AC-<n> (AC-1, AC-2, ...) — do not reorder IDs across re-plans; new ACs get fresh sequential IDs.
-- Every AC row MUST carry a Disposition value from the enum: new | updated | unchanged.
-- Each Outcome MUST be a single user-visible behaviour — no DB assertions, no HTTP status codes, no internal implementation details.
-- Cite proposed feature file paths under tests/features/** so Phase 8 can scaffold matching scenarios.`;
-
-/**
- * Build the authoring context the host LLM (or the
- * `epic-plan-spec-author` Skill) needs to write the PRD and Tech Spec.
- *
- * Returns a plain JSON-serialisable object; never hits the network beyond the
- * provider call needed to load the Epic.
- *
- * `docsContext` is bounded by the planning-context budget (Epic #817 Story 9):
- * over-budget payloads downgrade to a summary representation with headings +
- * bounded excerpts. Pass `{ fullContext: true }` (CLI: `--full-context`) to
- * restore the unbounded full-body envelope. The Epic body itself is always
- * subject to the same budget so a sprawling Epic narrative cannot bypass the
- * cap by riding on top of `docsContext`.
- */
-export async function buildAuthoringContext(
-  epicId,
-  provider,
-  settings = {},
-  opts = {},
-) {
-  const epic = await provider.getEpic(epicId);
-  if (!epic) {
-    throw new Error(`Epic #${epicId} not found.`);
-  }
-
-  const planningLimits = getLimits(settings).planningContext;
-  const { fullContext = false } = opts;
-
-  const docsContext = await buildDocsContext(settings, planningLimits, {
-    fullContext,
-  });
-
-  const epicBody = applyBudget(
-    [{ path: `epic-${epic.id}.md`, content: epic.body ?? '' }],
-    planningLimits,
-    { fullContext },
-  );
-
-  // Story #2094 Task #2103 — verify the project's BDD runner pending-tag
-  // support so the acceptance-spec body can record either the verified tag
-  // (features-first ordering) or "fallback: dependencies-first ordering"
-  // when no supported runner is present.
-  const bddRunner = await verifyBddRunnerPendingTag({ cwd: PROJECT_ROOT });
-
-  // Story #2637 — index existing BDD scenarios so the Acceptance Engineer
-  // step can annotate planned ACs with matches from the project's
-  // `.feature` files. Empty array when the project has not adopted BDD;
-  // the scanner is best-effort and never throws on filesystem errors.
-  let bddScenarios = [];
-  try {
-    const featureRoots = resolveFeatureRoots({ cwd: PROJECT_ROOT });
-    bddScenarios = scanBddScenarios({ featureRoots });
-  } catch (err) {
-    Logger.warn(`[epic-plan-spec] BDD scenario scan skipped: ${err.message}`);
-  }
-
-  // Story #2557 — memory-freshness pre-flight runs BEFORE the prior-feedback
-  // fetch so the planner sees a deduplicated, currently-actionable memory
-  // store. The scanner is best-effort: missing memory dir or gh-CLI failures
-  // land in `memoryFreshness.errors[]` and never throw.
-  const githubCfg = opts.github ?? null;
-  const memoryDir = resolveMemoryDir({ github: githubCfg });
-  const memoryFreshness = await scanMemoryFreshness({
-    memoryDir,
-    owner: githubCfg?.owner,
-    repo: githubCfg?.repo,
-    projectRoot: PROJECT_ROOT,
-  });
-
-  // Story #2554 — surface open meta feedback issues to the planner so retro
-  // signals are routed into durable substrates rather than lost in chat.
-  // The fetcher is best-effort: missing owner/repo or gh-CLI failures land
-  // in `errors[]` and never throw.
-  const priorFeedback = await fetchPriorFeedback({
-    owner: githubCfg?.owner,
-    repo: githubCfg?.repo,
-  });
-
-  // Story #2634 — codebase snapshot. Generates a bounded structural view
-  // of the consumer repo (file tree + package surface + recent activity
-  // + optional export signatures at the `medium` tier) so the Architect
-  // can prefer real module names over doc-only ones. The check is
-  // best-effort: any git/filesystem error degrades to an empty snapshot
-  // so Phase 7 stays non-blocking.
-  let codebaseSnapshot = null;
-  try {
-    codebaseSnapshot = buildCodebaseSnapshot({
-      cwd: PROJECT_ROOT,
-      tier: settings?.planning?.codebaseSnapshot?.tier,
-      include: settings?.planning?.codebaseSnapshot?.include,
-      exclude: settings?.planning?.codebaseSnapshot?.exclude,
-      recentCommitWindow:
-        settings?.planning?.codebaseSnapshot?.recentCommitWindow,
-    });
-  } catch (err) {
-    Logger.warn(`[epic-plan-spec] codebase snapshot skipped: ${err.message}`);
-  }
-
-  // Story #2791 — deterministic planning-risk envelope for gate routing
-  // and acceptance disposition. Pure classification over Epic metadata;
-  // no GitHub mutation on the emit-context path.
-  const planningRisk = classifyPlanningRisk({
-    title: epic.title,
-    body: epic.body ?? '',
-    labels: epic.labels ?? [],
-  });
-
-  return {
-    epic: {
-      id: epic.id,
-      title: epic.title,
-      body: epicBody.mode === 'full' ? epic.body : null,
-      bodySummary: epicBody.mode === 'summary' ? epicBody.items[0] : null,
-      linkedIssues: epic.linkedIssues ?? { prd: null, techSpec: null },
-    },
-    docsContext,
-    codebaseSnapshot,
-    systemPrompts: {
-      prd: PRD_SYSTEM_PROMPT,
-      techSpec: TECH_SPEC_SYSTEM_PROMPT,
-      acceptanceSpec: ACCEPTANCE_SPEC_SYSTEM_PROMPT,
-    },
-    bddRunner,
-    bddScenarios,
-    memoryFreshness,
-    priorFeedback,
-    planningRisk,
-  };
-}
-
-/**
- * Resolve the per-project memory directory used by the memory-freshness
- * pre-flight (Story #2557 / Epic #2547).
- *
- * Resolution order:
- *   1. `MANDREL_MEMORY_DIR` environment variable (test seam and operator
- *      override).
- *   2. `~/.claude/projects/<repo>/memory/` — the standard Claude Code
- *      memory substrate path, scoped by the configured GitHub repo so each
- *      consumer project gets its own memory pool.
- *   3. `null` when neither is resolvable. The scanner tolerates a missing
- *      `memoryDir` and surfaces a single `errors[]` entry.
- *
- * @param {{ github?: { owner?: string, repo?: string }|null }} opts
- * @returns {string|null}
- */
-export function resolveMemoryDir({ github } = {}) {
-  if (
-    typeof process.env.MANDREL_MEMORY_DIR === 'string' &&
-    process.env.MANDREL_MEMORY_DIR.length > 0
-  ) {
-    return process.env.MANDREL_MEMORY_DIR;
-  }
-  const repo = github?.repo;
-  if (typeof repo !== 'string' || repo.length === 0) return null;
-  return path.join(os.homedir(), '.claude', 'projects', repo, 'memory');
-}
-
-/**
- * Resolve whether Phase 7 should persist an acceptance-spec ticket or apply
- * the existing `acceptance::n-a` waiver from {@link classifyPlanningRisk}.
- *
- * @param {{ title?: string, body?: string, labels?: string[] }} epic
- * @param {string|null} acceptanceSpecContent
- * @returns {{ planningRisk: import('./lib/orchestration/planning-risk.js').PlanningRiskEnvelope, wantsAcceptanceSpec: boolean, applyAcceptanceWaiver: boolean }}
- */
-export function resolveAcceptancePersistence(epic, acceptanceSpecContent) {
-  const planningRisk = classifyPlanningRisk({
-    title: epic.title,
-    body: epic.body ?? '',
-    labels: epic.labels ?? [],
-  });
-
-  const hasAcceptanceContent =
-    typeof acceptanceSpecContent === 'string' &&
-    acceptanceSpecContent.trim() !== '';
-
-  if (planningRisk.acceptanceDisposition === 'not-applicable') {
-    return {
-      planningRisk,
-      wantsAcceptanceSpec: false,
-      applyAcceptanceWaiver: true,
-    };
-  }
-
-  return {
-    planningRisk,
-    wantsAcceptanceSpec: hasAcceptanceContent,
-    applyAcceptanceWaiver: false,
-  };
-}
-
-/**
- * Persist the host-authored PRD and Tech Spec under the Epic.
- *
- * Heals any prior planning artifacts (PRD / Tech Spec issues, "Planning
- * Artifacts" body section, lifecycle labels) before writing the new issues.
- * Idempotent against partial state: when the Epic already has a PRD but no
- * Tech Spec, the existing PRD is reused. Pass `force: true` to force a full
- * regeneration (close prior PRD/Tech Spec and re-create both).
- */
-export async function planEpic(
-  epicId,
-  provider,
-  { prdContent, techSpecContent, acceptanceSpecContent = null },
-  _settings = {},
-  { force = false } = {},
-) {
-  if (typeof prdContent !== 'string' || prdContent.trim() === '') {
-    throw new Error(
-      '[Epic Planner] prdContent is required and must be non-empty.',
-    );
-  }
-  if (typeof techSpecContent !== 'string' || techSpecContent.trim() === '') {
-    throw new Error(
-      '[Epic Planner] techSpecContent is required and must be non-empty.',
-    );
-  }
-  if (
-    acceptanceSpecContent !== null &&
-    (typeof acceptanceSpecContent !== 'string' ||
-      acceptanceSpecContent.trim() === '')
-  ) {
-    throw new Error(
-      '[Epic Planner] acceptanceSpecContent, when provided, must be a non-empty string.',
-    );
-  }
-
-  Logger.info(`[Epic Planner] Fetching Epic #${epicId}...`);
-  const epic = await provider.getEpic(epicId);
-
-  if (!epic) {
-    throw new Error(`Epic #${epicId} not found.`);
-  }
-
-  const stateManager = new PlanningStateManager(provider);
-  await stateManager.healAndCleanupArtifacts(epic, force);
-
-  const { planningRisk, wantsAcceptanceSpec, applyAcceptanceWaiver } =
-    resolveAcceptancePersistence(epic, acceptanceSpecContent);
-
-  Logger.info(
-    `[Epic Planner] Acceptance disposition: ${planningRisk.acceptanceDisposition}` +
-      (applyAcceptanceWaiver
-        ? ` — applying ${ACCEPTANCE_NA} waiver (no acceptance-spec ticket).`
-        : wantsAcceptanceSpec
-          ? ' — persisting context::acceptance-spec.'
-          : ' — no acceptance-spec content supplied.'),
-  );
-
-  // M-8: Resumable planning — if all artifacts exist, abort to prevent dupes.
-  const hasPrd = Boolean(epic.linkedIssues?.prd);
-  const hasTechSpec = Boolean(epic.linkedIssues?.techSpec);
-  const hasAcceptanceSpec = Boolean(epic.linkedIssues?.acceptanceSpec);
-  const allLinked =
-    hasPrd && hasTechSpec && (wantsAcceptanceSpec ? hasAcceptanceSpec : true);
-  if (!force && allLinked) {
-    Logger.warn(
-      `[Epic Planner] Epic #${epicId} already has all requested planning artifacts. Aborting to prevent duplicates. Use --force to re-plan.`,
-    );
-    return;
-  }
-  const existingPrdId = force ? null : (epic.linkedIssues?.prd ?? null);
-  const existingTechSpecId = force
-    ? null
-    : (epic.linkedIssues?.techSpec ?? null);
-  const existingAcceptanceSpecId = force
-    ? null
-    : (epic.linkedIssues?.acceptanceSpec ?? null);
-
-  let prdId;
-  if (existingPrdId) {
-    Logger.info(
-      `[Epic Planner] Reusing existing PRD #${existingPrdId}. Skipping PRD creation.`,
-    );
-    prdId = existingPrdId;
-  } else {
-    Logger.info(`[Epic Planner] Creating PRD issue for "${epic.title}"...`);
-    const prdTicket = await provider.createTicket(epicId, {
-      title: `[PRD] ${epic.title}`,
-      body: prdContent,
-      labels: ['context::prd'],
-      dependencies: [],
-    });
-    Logger.info(
-      `[Epic Planner] Created PRD Issue #${prdTicket.id} (${prdTicket.url})`,
-    );
-    prdId = prdTicket.id;
-  }
-
-  let techSpecId;
-  if (existingTechSpecId) {
-    Logger.info(
-      `[Epic Planner] Reusing existing Tech Spec #${existingTechSpecId}. Skipping Tech Spec creation.`,
-    );
-    techSpecId = existingTechSpecId;
-  } else {
-    Logger.info(
-      `[Epic Planner] Creating Tech Spec issue linking to PRD #${prdId}...`,
-    );
-    const techSpecTicket = await provider.createTicket(epicId, {
-      title: `[Tech Spec] ${epic.title}`,
-      body: techSpecContent,
-      labels: ['context::tech-spec'],
-      dependencies: [prdId],
-    });
-    Logger.info(
-      `[Epic Planner] Created Tech Spec Issue #${techSpecTicket.id} (${techSpecTicket.url})`,
-    );
-    techSpecId = techSpecTicket.id;
-  }
-
-  let acceptanceSpecId = null;
-  if (wantsAcceptanceSpec) {
-    if (existingAcceptanceSpecId) {
-      Logger.info(
-        `[Epic Planner] Reusing existing Acceptance Spec #${existingAcceptanceSpecId}. Skipping Acceptance Spec creation.`,
-      );
-      acceptanceSpecId = existingAcceptanceSpecId;
-    } else {
-      Logger.info(
-        `[Epic Planner] Creating Acceptance Spec issue linking to Tech Spec #${techSpecId}...`,
-      );
-      const acceptanceTicket = await provider.createTicket(epicId, {
-        title: `[Acceptance Spec] ${epic.title}`,
-        body: acceptanceSpecContent,
-        labels: ['context::acceptance-spec'],
-        dependencies: [techSpecId],
-      });
-      Logger.info(
-        `[Epic Planner] Created Acceptance Spec Issue #${acceptanceTicket.id} (${acceptanceTicket.url})`,
-      );
-      acceptanceSpecId = acceptanceTicket.id;
-    }
-  }
-
-  Logger.info(
-    `[Epic Planner] Updating Epic #${epicId} with linked documents...`,
-  );
-
-  // Format exactly so the issue-link-parser regexes still catch each line.
-  // The parser is the source of truth for which prefixes are accepted; we
-  // emit the canonical "PRD: #N" / "Tech Spec: #N" / "Acceptance Spec: #N"
-  // shape so cascade-close (epic-deliver-finalize) and the Phase 2
-  // decomposer-context picker both see the third link.
-  const artifactLines = [
-    `- [ ] PRD: #${prdId}`,
-    `- [ ] Tech Spec: #${techSpecId}`,
-  ];
-  if (acceptanceSpecId !== null) {
-    artifactLines.push(`- [ ] Acceptance Spec: #${acceptanceSpecId}`);
-  }
-  const appendBody = `\n\n## Planning Artifacts\n${artifactLines.join('\n')}\n`;
-  const newBody = epic.body + appendBody;
-
-  /** @type {{ add?: string[], remove?: string[] }} */
-  const labelMutations = {};
-  if (applyAcceptanceWaiver) {
-    labelMutations.add = [ACCEPTANCE_NA];
-  } else if (
-    wantsAcceptanceSpec &&
-    (epic.labels ?? []).includes(ACCEPTANCE_NA)
-  ) {
-    labelMutations.remove = [ACCEPTANCE_NA];
-  }
-
-  await provider.updateTicket(epicId, {
-    body: newBody,
-    ...(labelMutations.add || labelMutations.remove
-      ? { labels: labelMutations }
-      : {}),
-  });
-
-  Logger.info(`[Epic Planner] Epic #${epicId} updated successfully.`);
-  Logger.info(`[Epic Planner] Planning pipeline complete!`);
-}
-
-/**
- * Runs `sweepStaleStoryWorktrees` when a ticketing `provider` is available
- * (normal CLI boot): drains `.pending-cleanup.json` with Windows escalation,
- * then reaps registered worktrees for done/closed stories. When `provider` is
- * omitted (unit tests), runs `forceDrainPendingCleanup` on the manifest only.
- *
- * Uses `delivery.worktreeIsolation.root` when present; defaults to
- * `.worktrees`.
- *
- * Non-blocking: stuck entries stay in the manifest; plan execution continues.
- *
- * Exposed for integration tests.
- *
- * @param {{ repoRoot?: string, config?: object, provider?: object, git?: object, logger?: object, fsRm?: function }} [opts]
- * @returns {Promise<object>} Sweep/drain summary with legacy `drained` / `persistent` / `remaining` aliases for callers.
- */
-export async function drainPendingCleanupAtBoot(opts = {}) {
-  const repoRoot = opts.repoRoot ?? PROJECT_ROOT;
-  const config = opts.config;
-  const worktreeRoot = path.join(
-    repoRoot,
-    config?.delivery?.worktreeIsolation?.root ?? '.worktrees',
-  );
-  const git = opts.git ?? gitUtils;
-  const logger = opts.logger ?? console;
-  const fsRm = opts.fsRm;
-  const provider = opts.provider;
-
-  // legacyExtras adds `drained`/`persistent`/`remaining` aliases consumed by
-  // drain-pending-cleanup.js, epic-plan-decompose.js, plan-runner/worktree-sweep.js,
-  // and tests/epic-plan-spec-drain.test.js (Epic #990 Story #1006 triage).
-  const legacyExtras = (base) => {
-    const remaining =
-      (base.persistentPending?.length ?? base.persistent?.length ?? 0) +
-      (base.stillPending?.length ?? 0);
-    const drained = base.drainedPending ?? base.drained ?? [];
-    const persistent = base.persistentPending ?? base.persistent ?? [];
-    return {
-      ...base,
-      remaining,
-      drained,
-      persistent,
-    };
-  };
-
-  if (provider?.getTicket) {
-    const sweep = await sweepStaleStoryWorktrees({
-      provider,
-      repoRoot,
-      git,
-      logger,
-      worktreeRoot,
-      fsRm,
-    });
-    const remaining =
-      (sweep.persistentPending?.length ?? 0) +
-      (sweep.stillPending?.length ?? 0);
-    logger.info?.(
-      `[epic-plan-spec] worktree sweep: reaped=${sweep.reaped.length} drainedPending=${sweep.drainedPending?.length ?? 0} remaining=${remaining}`,
-    );
-    return legacyExtras(sweep);
-  }
-
-  const before = readManifest(worktreeRoot).length;
-  if (before === 0) {
-    return legacyExtras({
-      reaped: [],
-      skipped: [],
-      drainedPending: [],
-      persistentPending: [],
-      stillPending: [],
-    });
-  }
-  const result = await forceDrainPendingCleanup({
-    repoRoot,
-    worktreeRoot,
-    git,
-    fsRm,
-    logger,
-  });
-  const remaining =
-    (result.persistent?.length ?? 0) + (result.stillPending?.length ?? 0);
-  logger.info?.(
-    `[epic-plan-spec] pending-cleanup drain: reaped=${result.drained?.length ?? 0} remaining=${remaining}`,
-  );
-  return legacyExtras({
-    reaped: [],
-    skipped: [],
-    drainedPending: result.drained,
-    persistentPending: result.persistent,
-    stillPending: result.stillPending,
-    escalated: result.escalated,
-    killedPids: result.killedPids,
-    noHolders: result.noHolders,
-    drainedDetails: result.drainedDetails,
-    persistentDetails: result.persistentDetails,
-    stillPendingDetails: result.stillPendingDetails,
-  });
-}
-
-async function setEpicLabel(provider, epicId, targetLabel) {
-  const planningLabels = [AGENT_LABELS.REVIEW_SPEC, AGENT_LABELS.READY];
-  await provider.updateTicket(epicId, {
-    labels: {
-      add: [targetLabel],
-      remove: planningLabels.filter((l) => l !== targetLabel),
-    },
-  });
-}
-
-/**
- * Run the Tech Spec freshness check against the current codebase and emit
- * the advisory artefacts.
- *
- * Two side effects, both best-effort and non-blocking:
- *   1. Write the full `{ stale, fresh, ambiguous }` report to
- *      `<tempRoot>/epic-<id>-spec-freshness.json` so downstream tooling
- *      (the `--code-freshness` health check on the roadmap, or an
- *      operator inspecting the run) can read it without re-probing git.
- *   2. When `stale.length > 0`, upsert a `spec-freshness` structured
- *      comment on the Tech Spec issue listing each suspect citation.
- *      The comment is upserted, so re-running spec re-renders the same
- *      comment in place rather than spamming the issue.
- *
- * Caller signals are aggressive in containment — any error in this path
- * is downgraded to a warning. Phase 7 must not fail because a doc-author
- * cited a path the validator can't probe (e.g. shallow clone, base ref
- * not fetched). The non-blocking contract is the load-bearing AC.
- *
- * @param {object} opts
- * @param {number} opts.epicId
- * @param {number|null} opts.techSpecId
- * @param {string} opts.techSpecContent
- * @param {string} opts.baseBranchRef
- * @param {string} opts.tempRoot
- * @param {import('./lib/ITicketingProvider.js').ITicketingProvider} opts.provider
- * @param {Function} [opts.validator] - Testing seam (defaults to validateSpecFreshness).
- * @param {Function} [opts.commentUpserter] - Testing seam (defaults to upsertStructuredComment).
- * @param {Function} [opts.fileWriter] - Testing seam (defaults to fs writeFile/mkdir wrapper).
- * @returns {Promise<{ stale: number, ambiguous: number, fresh: number, reportPath: string|null, commentPosted: boolean }>}
- */
-export async function runSpecFreshnessCheck({
-  epicId,
-  techSpecId,
-  techSpecContent,
-  baseBranchRef,
-  tempRoot,
-  provider,
-  validator = validateSpecFreshness,
-  commentUpserter = upsertStructuredComment,
-  fileWriter = async (filePath, body) => {
-    await mkdir(path.dirname(filePath), { recursive: true });
-    await writeFile(filePath, body, 'utf8');
-  },
-}) {
-  try {
-    const report = validator(techSpecContent, { baseBranchRef });
-    const reportPath = path.join(
-      tempRoot,
-      `epic-${epicId}-spec-freshness.json`,
-    );
-    const payload = {
-      epicId,
-      techSpecId,
-      baseBranchRef,
-      generatedAt: new Date().toISOString(),
-      summary: {
-        stale: report.stale.length,
-        ambiguous: report.ambiguous.length,
-        fresh: report.fresh.length,
-      },
-      ...report,
-    };
-    await fileWriter(reportPath, `${JSON.stringify(payload, null, 2)}\n`);
-
-    let commentPosted = false;
-    if (report.stale.length > 0 && Number.isFinite(techSpecId)) {
-      const body = renderSpecFreshnessComment(report, {
-        baseBranchRef,
-        techSpecId,
-        epicId,
-      });
-      await commentUpserter(provider, techSpecId, 'spec-freshness', body);
-      commentPosted = true;
-    }
-
-    if (report.stale.length > 0 || report.ambiguous.length > 0) {
-      Logger.warn(
-        `[epic-plan-spec] Tech Spec freshness: ${report.stale.length} stale, ${report.ambiguous.length} ambiguous, ${report.fresh.length} fresh against ${baseBranchRef}. Report: ${reportPath}.`,
-      );
-    } else {
-      Logger.info(
-        `[epic-plan-spec] Tech Spec freshness: ${report.fresh.length} fresh references against ${baseBranchRef}. No drift detected.`,
-      );
-    }
-
-    return {
-      stale: report.stale.length,
-      ambiguous: report.ambiguous.length,
-      fresh: report.fresh.length,
-      reportPath,
-      commentPosted,
-    };
-  } catch (err) {
-    Logger.warn(
-      `[epic-plan-spec] Tech Spec freshness check skipped: ${err.message}`,
-    );
-    return {
-      stale: 0,
-      ambiguous: 0,
-      fresh: 0,
-      reportPath: null,
-      commentPosted: false,
-      error: err.message,
-    };
-  }
-}
-
-/**
- * Execute the spec phase end to end.
- *
- * @param {number} epicId
- * @param {import('./lib/ITicketingProvider.js').ITicketingProvider} provider
- * @param {{ prdContent: string, techSpecContent: string }} artifacts
- * @param {object} settings
- * @param {{ force?: boolean, forceReview?: boolean, snapshotFork?: typeof forkAndCommitEpicSnapshot }} [opts]
- * @returns {Promise<{ epicId: number, prdId: number|null, techSpecId: number|null, checkpoint: object, planningRisk: import('./lib/orchestration/planning-risk.js').PlanningRiskEnvelope, reviewRouting: import('./lib/orchestration/plan-review-routing.js').ReviewRoutingEnvelope }>}
- */
-export async function runSpecPhase(
-  epicId,
-  provider,
-  { prdContent, techSpecContent, acceptanceSpecContent = null },
-  settings = {},
-  { force = false, forceReview = false } = {},
-) {
-  const epic = await provider.getEpic(epicId);
-  if (!epic) {
-    throw new Error(`[epic-plan-spec] Epic #${epicId} not found.`);
-  }
-  if (!epic.labels?.includes(TYPE_LABELS.EPIC)) {
-    throw new Error(
-      `[epic-plan-spec] Ticket #${epicId} is not a ${TYPE_LABELS.EPIC}.`,
-    );
-  }
-
-  await initializePlanState({ provider, epicId });
-  await setPlanPhase({
-    provider,
-    epicId,
-    nextPhase: PLAN_PHASES.PLANNING,
-  });
-
-  await planEpic(
-    epicId,
-    provider,
-    { prdContent, techSpecContent, acceptanceSpecContent },
-    settings,
-    {
-      force,
-    },
-  );
-
-  const afterPlan = await provider.getEpic(epicId);
-  const prdId = afterPlan.linkedIssues?.prd ?? null;
-  const techSpecId = afterPlan.linkedIssues?.techSpec ?? null;
-  const acceptanceSpecId = afterPlan.linkedIssues?.acceptanceSpec ?? null;
-
-  // Story #2635 — cross-validate the authored Tech Spec body against the
-  // base branch and surface any stale path-shaped references on the Tech
-  // Spec issue. Non-blocking: a missing base ref, an unreadable temp
-  // directory, or a provider failure downgrades to a warning so Phase 7
-  // never fails on the advisory check.
-  const baseBranchRef = settings?.baseBranch ?? 'main';
-  const tempRoot = path.resolve(
-    PROJECT_ROOT,
-    settings?.paths?.tempRoot ?? 'temp',
-  );
-  const freshness = await runSpecFreshnessCheck({
-    epicId,
-    techSpecId,
-    techSpecContent,
-    baseBranchRef,
-    tempRoot,
-    provider,
-  });
-
-  // Story #1585 (Epic #1471): the baseline-snapshot fork was previously
-  // performed here at plan-time. It now runs at first-story-init time
-  // inside `lib/story-init/branch-initializer.js#bootstrapWorktree` so
-  // `/epic-plan` remains git-state-free. `forkAndCommitEpicSnapshot` and
-  // `forkMainToEpic` remain exported for that caller.
-
-  const planningRisk = classifyPlanningRisk({
-    title: afterPlan.title,
-    body: afterPlan.body ?? '',
-    labels: afterPlan.labels ?? [],
-  });
-  const reviewRouting = resolveReviewRouting({ planningRisk, forceReview });
-
-  const currentState =
-    (await readPlanState({ provider, epicId })) ??
-    (await initializePlanState({ provider, epicId }));
-  const checkpoint = await writePlanState({
-    provider,
-    epicId,
-    state: {
-      ...currentState,
-      planningRisk,
-      reviewRouting: {
-        decision: reviewRouting.decision,
-        requiresStop: reviewRouting.requiresStop,
-        forceReviewApplied: reviewRouting.forceReviewApplied,
-      },
-      spec: {
-        ...currentState.spec,
-        prdId,
-        techSpecId,
-        acceptanceSpecId,
-        completedAt: new Date().toISOString(),
-      },
-    },
-  });
-
-  Logger.info(`[epic-plan-spec] Review routing: ${reviewRouting.decision}.`);
-  Logger.info(`[epic-plan-spec] ${reviewRouting.operatorMessage}`);
-
-  Logger.info(
-    `[epic-plan-spec] Flipping Epic #${epicId} to ${AGENT_LABELS.REVIEW_SPEC}...`,
-  );
-  await setEpicLabel(provider, epicId, AGENT_LABELS.REVIEW_SPEC);
-  await setPlanPhase({
-    provider,
-    epicId,
-    nextPhase: PLAN_PHASES.REVIEW_SPEC,
-  });
-
-  const cleanup = await cleanupPhaseTempFiles({ phase: 'spec', epicId });
-
-  const acceptanceSummary =
-    acceptanceSpecId !== null ? `, Acceptance Spec #${acceptanceSpecId}` : '';
-  const freshnessSummary =
-    freshness.stale > 0 || freshness.ambiguous > 0
-      ? ` ⚠️ Spec freshness: ${freshness.stale} stale / ${freshness.ambiguous} ambiguous reference(s) — see ${freshness.reportPath ?? 'report'}.`
-      : '';
-  Logger.info(
-    `[epic-plan-spec] ✅ Spec phase complete for Epic #${epicId}. PRD #${prdId}, Tech Spec #${techSpecId}${acceptanceSummary}.${freshnessSummary}`,
-  );
-  if (cleanup.deleted.length > 0) {
-    Logger.info(
-      `[epic-plan-spec] 🧹 Cleaned up ${cleanup.deleted.length} temp file(s).`,
-    );
-  }
-
-  return {
-    epicId,
-    prdId,
-    techSpecId,
-    acceptanceSpecId,
-    checkpoint,
-    cleanup,
-    freshness,
-    planningRisk,
-    reviewRouting,
-  };
-}
-
-/* node:coverage ignore next */
 async function main() {
   const { values } = parseArgs({
     options: {
@@ -918,8 +116,8 @@ async function main() {
   try {
     config = resolveConfig();
     // `settings` retains the legacy bag shape used by buildAuthoringContext
-    // and friends in this file: `{ baseBranch, paths, planning, ... }`.
-    // Build it from the canonical blocks rather than the legacy shim.
+    // and friends: `{ baseBranch, paths, planning, ... }`. Build it from the
+    // canonical blocks rather than the legacy shim.
     settings = {
       baseBranch: config.project?.baseBranch,
       paths: config.project?.paths,
